fix(#332): unify cosign path probe between doctor and fetch-image
External user @garyoakidoki's #332 report: \`aegis-boot doctor\` sees
cosign at \`/usr/local/bin/cosign\` and reports \`[✓ PASS]\`, but the
very next command \`aegis-boot fetch-image\` says "cosign not on PATH"
and silently falls through to an unverified download.

Root cause: two different probes.

- \`doctor::check_cosign_optional\` used \`which("cosign")\` — a file-
  existence check against \`\$PATH\` + (post-#331) sbin fallbacks.
- \`fetch_image::cosign_on_path\` shelled out to \`cosign --version\`
  and checked the exit code. That probe returns false for reasons
  beyond "binary missing" — a broken binary, a non-zero exit from a
  cosign version that's fussy about its config dir, a transparency-
  log lookup that times out. The user's particular cosign exits
  non-zero on \`--version\` for whatever reason, so fetch-image read
  it as absent while doctor correctly found it.

Fix: move \`which\` / \`which_in\` into a new shared \`src/cmd_path.rs\`
module so every aegis-boot surface asking "is this command
available?" uses the same probe. Both \`doctor\` and \`fetch-image\`
import from it. \`cosign_on_path\` becomes a one-liner that delegates.

No behavior change for the happy path (cosign in \$PATH and not
broken → both probes return true). Behavior change for the broken/
fussy-version case: fetch-image now attempts the verify-blob call,
which will fail with the real cosign error — much more actionable
than "cosign not on PATH". For the "cosign genuinely missing" case:
same warning as before, same graceful-degrade to unverified download.

Tests moved along with \`which_in\` to \`cmd_path::tests\`. The doctor
tests that exercised the function via an imported name are unaffected
since the re-export at the module boundary is transparent.
